added lambok support to model class

+++ src/main/java/com/example/spring_rest_api/services/StudentService.java
@@ -4,7 +4,7 @@
  */
 package com.example.spring_rest_api.services;
 
-import com.example.spring_rest_api.model.Student;
+import com.example.spring_rest_api.models.Student;
 import com.example.spring_rest_api.repository.StudentRepository;
 import java.util.List;
 import java.util.Optional;
@@ -32,19 +32,47 @@ public class StudentService {
     }
 
     public void createStudent(Student student) {
-        studentRepository.save(student);
+        Optional<Student> studentOptional = studentRepository.findStudentByEmail(student.getEmail());
+        if (studentOptional.isPresent()) {
+            throw new IllegalStateException("failed to create new student because email already exist");
+        } else {
+            studentRepository.save(student);
+        }
     }
 
-    public void editStudent(Student student) {
-      Optional<Student> student1 = studentRepository.findById(student.getId());
-      student1.get().setName(student.getName());
-      student1.get().setEmail(student.getEmail());
-      student1.get().setDob(student.getDob());
-      studentRepository.save(student1.get());
+    public void editStudent(Long id, String name, String email) {
+        Optional<Student> studenOptional = studentRepository.findById(id);
+
+        if (!studenOptional.isPresent()) {
+            throw new IllegalStateException("failed to update student with id " + id + " because they do not exist");
+        } else {
+
+            if (email != null && email.length() > 0) {
+                if (studentRepository.findStudentByEmail(email).isPresent()) {
+                    throw new IllegalStateException("failed to update email because email already exist");
+                } else {
+                    studenOptional.get().setEmail(email);
+                }
+            }
+
+            if (name != null && name.length() > 0) {
+                if (studenOptional.get().getName().equals(name)) {
+                    throw new IllegalStateException("failed to update name because name has not changed");
+                } else {
+                    studenOptional.get().setName(name);
+                }
+            }
+             studentRepository.save(studenOptional.get());
+        }
     }
 
     public void deleteStudent(Long id) {
-        studentRepository.delete(studentRepository.findById(id).get());
+        Optional<Student> studentOptional = studentRepository.findById(id);
+        if (!studentOptional.isPresent()) {
+            throw new IllegalStateException("failed to delete student because student with id " + id + " does not exist");
+        } else {
+            studentRepository.delete(studentRepository.findById(id).get());
+        }
     }
 
     public Optional<Student> getStudentByEmail(String email) {
@@ -55,4 +83,3 @@ public class StudentService {
         return studentRepository.findStudentByName(name);
     }
 }
-
